Fix show continuity on playout end
This has involved some not-ideal level bodging.

The Brave image source appears to just show funky colours when it's put
into a mixer. The way to fix it whilst still using an image source is to
remove it from the mix and put it back and it magically corrects itself.
Can't do that as an approach since the user will still see funky
colours.

The current fix converts the card images into two second long videos and
creates them in Brave as URI looping inputs instead which appear to be
more stabler and no longer shows funky colours.

There are still issues. Firstly, Brave shows darkness between cuts,
might be able to fix it using the overlay function. Secondly, the
URI looping input causes Brave to fetch the video for every loop.
There is an attempt to address, but only works in the "off-air"
state, where it will pause the input 500ms after the new card
input being played.